backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
gunicorn==22.0.0